[AMD-SMI] Stop leaking ualoe_lib.h through public install headers
The public-facing impl/amd_smi_gpu_device.h pulled in ualoe_lib/ualoe_lib.h
via extern "C", then immediately contradicted that include with a local
'typedef int ualoe_handle_t;'. This forced the vendored ualoe_lib headers
to be installed under /opt/rocm/include/ualoe_lib/ to keep downstream
consumers compiling.

Drop the include from the public header (keep only the forward typedef),
move the ctors/dtor out-of-line into amd_smi_gpu_device.cc where the real
ualoe_lib.h is included, and remove the now-unneeded install(DIRECTORY
ualoe_lib) rule.
